bucket-A2: extract enhanced_node into coin-agnostic template (SAFE-ADDITIVE)
Lifts the EnhancedC2PoolNode implementation out of the compiled
src/c2pool/node/enhanced_node.cpp into a header-only, coin-agnostic
template EnhancedC2PoolNodeT<Config, ShareChain, Peer, ShareType> in the
new enhanced_node_impl.hpp. enhanced_node.hpp becomes a thin backward-compat
header that pins the template to LTC types (EnhancedC2PoolNode =
EnhancedC2PoolNodeT<ltc::...>), so existing callers compile unchanged.

c2pool_node_enhanced becomes an INTERFACE (header-only) library; the two
executables that consume it (c2pool, c2pool_enhanced) already link `ltc`
directly, so dropping `ltc` from the library's transitive deps is a no-op
for the LTC build.

Pure refactor — LTC behavior identical. The Dash specialization
(src/impl/dash/enhanced_node.hpp) is deferred to Bucket B with the rest of
src/impl/dash/, since master has no dash impl directory yet and that shim
references not-yet-landed dash headers.

Cherry-picked from dash-spv-embedded HEAD 8249dc28.

4 files, +251 / -368:
- new      src/c2pool/node/enhanced_node_impl.hpp (template)
- modified src/c2pool/node/enhanced_node.hpp        (LTC alias shim)
- deleted  src/c2pool/node/enhanced_node.cpp        (impl moved to template)
- modified src/c2pool/CMakeLists.txt                (library -> INTERFACE)
